correct log_schema
